requirements.txt
Flask==1.1.1
line-bot-sdk==1.15.0
matplotlib==3.5.1
